3 Arranged subdivision files of styled-components for FC CarPhoto
General.styled.ts
CarPhoto.styled.ts
CardPhotoPlates.styled.ts
