+++ src/components/CarPhoto/CarPhoto.tsx
@@ -1,5 +1,6 @@
 import React, { FC } from 'react';
 import * as card from './CarPhoto.styled';
+import * as plate from './CardPhotoPlate';
 
 interface CarPhotoProps {}
 
@@ -8,9 +9,9 @@ const CarPhoto: FC<CarPhotoProps> = () => (
    <card.CarPhotoImg src={require('./img/car_icon.jpg')}/>
    <card.CarPhotoInner _direction='column'>
       <card.ContentTop>
-         
+
       </card.ContentTop>
-      <card.ContentBottom>
+      <card.ContentBottom _direction='column'>
 
       </card.ContentBottom>
    </card.CarPhotoInner>
